leitarniðurstöður, manyTomany

+++ src/main/java/is/hi/hbv501g/freyr/Freyr/Repositories/UserRepository.java
@@ -1,12 +1,16 @@
 package is.hi.hbv501g.freyr.Freyr.Repositories;
 
+import is.hi.hbv501g.freyr.Freyr.Entities.Recipe;
 import is.hi.hbv501g.freyr.Freyr.Entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     // ATH AÐFERÐIRNAR SEM ERU COMMENTAÐAR ÚT HÉR ER EKKI HÆGT AÐ HAFA Í REPOINU
     // ÞETTA ERU AÐFERÐIR SEM ÞARF AÐ ÚTFÆRA Í SERVICE IMPLEMENTATION SÉRSTAKLEGA
